adding generic 'fast' operator and mapper

+++ board.js
@@ -1,6 +1,6 @@
 function Board(w, h) {
     var bounds = new Rect(0, 0, w, h);
-    var array = new Int32Array(w * h);
+    var array = new Int8Array(w * h);
     this.checkRange = (x, y) => bounds.inside(x, y);
     this.get = (x, y) => array[y * w + x];
     this.set = (x, y, b) => array[y * w + x] = b;
